+++ README.md
@@ -22,7 +22,7 @@
   * *从不回头看`“爆炸”`*
 ## 我怎么才能使用SRA？急急急
 下载链接：[SRAv0.6](https://github.com/Shasnow/StarRailAssisant/releases)
-* 如果你是小白，只需要下载项目中的`rar`文件，*一切都为您准备妥当*，只需解压到您喜欢的位置，然后运行`.exe`即可！
+* 如果你是小白，只需要下载项目中的`zip`文件，*一切都为您准备妥当*，只需解压到您喜欢的位置，然后运行`.exe`即可！
 * **我想挑战一下自己！** 当然没问题。打包下载项目中所有文件，确保你的电脑中已经安装好了`Python`，并安装好了下面的库，如果没有，使用`pip`来安装它们。
   ```bash
   pip install opencv-python
