+++ data/版本信息.txt
@@ -1,0 +1,5 @@
+v0.7_beta
+
+#1
+
+`这仅是一个测试版本，请普通用户下载上一个版本`
